requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.9
llama-cpp-python==0.2.20
sentence-transformers==2.7.0
numpy>=1.26.0
celery==5.3.4
redis==5.0.1
reportlab==4.0.7
markdown==3.5.1
python-dotenv==1.0.0
django-filter==23.5
drf-yasg==1.21.7
Pillow==10.1.0
torch>=2.0.0
transformers>=4.30.0

